requirements.txt
aiohttp==3.8.1
aioredis==1.3.1
aiosignal==1.2.0
amqp==5.0.9
asgiref==3.4.1
async-timeout==4.0.2
attrs==21.4.0
autobahn==21.11.1
Automat==20.2.0
base58==2.1.1
billiard==3.6.4.0
bitarray==1.2.2
boto3==1.21.3
botocore==1.24.3
celery==5.2.3
certifi==2021.10.8
cffi==1.15.0
channels==3.0.4
channels-redis==3.2.0
charset-normalizer==2.0.9
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==36.0.1
cytoolz==0.11.2
daphne==3.0.2
Deprecated==1.2.13
dj-database-url==0.5.0
Django==3.2.10
django-cors-headers==3.10.1
django-countries==7.2.1
django-dotenv==1.4.2
django-environ==0.8.1
django-phonenumber-field==6.0.0
django-rest-swagger==2.2.0
django-storages==1.12.3
djangorestframework==3.13.0
djangorestframework-jwt==1.11.0
eth-abi==2.1.1
eth-account==0.5.7
eth-hash==0.3.2
eth-keyfile==0.5.1
eth-keys==0.3.4
eth-rlp==0.2.1
eth-typing==2.3.0
eth-utils==1.10.0
frozenlist==1.3.0
gunicorn==20.1.0
h2==3.2.0
hexbytes==0.2.2
hiredis==2.0.0
hpack==3.0.0
hyperframe==5.2.0
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
ipfshttpclient==0.8.0a2
itypes==1.2.0
Jinja2==3.0.3
jmespath==0.10.0
jsonschema==3.2.0
kombu==5.2.3
lru-dict==1.1.7
MarkupSafe==2.0.1
msgpack==1.0.3
multiaddr==0.0.9
multidict==6.0.2
netaddr==0.8.0
openapi-codec==1.3.2
packaging==21.3
parsimonious==0.8.1
phonenumbers==8.12.39
pika==1.2.0
pika-stubs==0.1.3
Pillow==8.4.0
priority==1.3.0
prompt-toolkit==3.0.24
protobuf==3.19.4
psycopg2-binary==2.9.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pycryptodome==3.14.1
PyJWT==1.7.1
pyOpenSSL==21.0.0
pyparsing==3.0.6
pyrsistent==0.18.1
python-dateutil==2.8.2
python-decouple==3.5
pytz==2021.3
redis==4.1.0
requests==2.26.0
rlp==2.0.1
s3transfer==0.5.1
service-identity==21.1.0
simplejson==3.17.6
six==1.16.0
sqlparse==0.4.2
toolz==0.11.2
Twisted==21.7.0
txaio==21.2.1
typing_extensions==4.0.1
uritemplate==4.1.1
urllib3==1.26.7
varint==1.0.2
vine==5.0.0
wcwidth==0.2.5
web3==5.28.0
websockets==9.1
wrapt==1.13.3
yarl==1.7.2
zope.interface==5.4.0
